Add hook on soft-deleting user, move EntityDto type to a separate file
