+++ src/models/common/entity-dto.ts
@@ -1,0 +1,3 @@
+export type EntityDto = {
+    id: string;
+};
