+++ src/models/user.model.ts
@@ -1,6 +1,8 @@
 import { db } from '../config/dbConfig';
 import { BuildOptions, DataTypes, Model, Sequelize } from 'sequelize';
-import { EntityDto, Group } from './group.model';
+import { Group } from './group.model';
+import { UserGroup } from './user-group.model';
+import { EntityDto } from './common/entity-dto';
 
 export type UserDto = {
     id: string;
@@ -58,6 +60,12 @@ export const User = UserFactory(db);
 
 User.belongsToMany(Group, {
     through: 'user-group'
+});
+// eslint-disable-next-line no-unused-vars,@typescript-eslint/no-unused-vars
+User.afterUpdate((user, options) => {
+    if (user.isDeleted === true) {
+        UserGroup.destroy({ where: { userId: user.id } });
+    }
 });
 Group.belongsToMany(User, {
     through: 'user-group'
