+++ src/services/base-service.ts
@@ -1,7 +1,7 @@
 import BaseRepository from '../data-access/base-repository';
 import { BuildOptions, Model } from 'sequelize';
 import { BaseServiceI } from './base-service-i';
-import { EntityDto } from '../models/group.model';
+import { EntityDto } from '../models/common/entity-dto';
 
 export default class BaseService<M extends Model, T extends typeof Model & {
     new(values?: any, options?: BuildOptions): M;
